# basic_auth/model_unexpected_error.go
/*
 * myapi
 *
 * HTTP server in GO with swagger codegen
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package node

type UnexpectedError struct {
	Code int32 `json:"code,omitempty"`
	Message string `json:"message,omitempty"`
}