# basic_auth_server/go/api_default.go
/*
 * myapi
 *
 * HTTP server in GO with swagger codegen
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package node_server


import (
	"net/http"
    "encoding/json"
)


func CheckHealth(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	w.WriteHeader(http.StatusOK)
    json.NewEncoder(w).Encode(`{"Health" : "is OK"}`)

}


func HelloUser(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	w.WriteHeader(http.StatusOK)
}